requirements.txt
blinker==1.9.0
click==8.1.8
Flask==2.3.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.2.2
greenlet==3.2.1
gunicorn==22.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
SQLAlchemy==2.0.40
typing_extensions==4.13.2
Werkzeug==3.0.1
WTForms==3.2.1
Flask-Migrate==4.0.7
Flask-Migrate==4.0.7
